PSR-4-ify countries.php
