+++ src/MLD/CountryData/AbstractJsonConverter.php
@@ -1,0 +1,19 @@
+<?php
+namespace MLD\CountryData;
+
+/**
+ * Class AbstractJsonConverter
+ */
+abstract class AbstractJsonConverter extends AbstractConverter {
+
+    /**
+     * Special case for empty arrays that should be encoded as empty JSON objects
+     */
+    protected function processEmptyArrays() {
+        array_walk($this->aCountries, function (&$aCountry) {
+            if (empty($aCountry['languages'])) {
+                $aCountry['languages'] = new \stdClass();
+            }
+        });
+    }
+}
