+++ src/MLD/CountryData/JsonConverter.php
@@ -1,0 +1,16 @@
+<?php
+namespace MLD\CountryData;
+
+/**
+ * Class JsonConverter
+ */
+class JsonConverter extends AbstractJsonConverter {
+
+    /**
+     * @return string minified JSON, one country per line
+     */
+    public function convert() {
+        $this->processEmptyArrays();
+        return preg_replace("@},{@", "},\n{", json_encode($this->aCountries) . "\n");
+    }
+}
